WIP: switch encoding unsigned type normalization to Uint* types
Using Java native types for unsigned values is wasteful especially
in case of uint32 and uint64. Furthermore it leads to uncertainty
between in16/uint8, int32/uint16 and int64/uint32.

This is a partial switch, where we retain java.lang types for
signed entities, but switch the unsigned ones.

+++ yang/yang-data-impl/src/test/java/org/opendaylight/yangtools/yang/data/impl/codecs/Uint32CodecStringTest.java
@@ -12,6 +12,7 @@ import static org.opendaylight.yangtools.yang.data.impl.codecs.TypeDefinitionAwa
 import static org.opendaylight.yangtools.yang.data.impl.codecs.TypeDefinitionAwareCodecTestHelper.getCodec;
 
 import org.junit.Test;
+import org.opendaylight.yangtools.yang.common.Uint32;
 import org.opendaylight.yangtools.yang.data.api.codec.Uint32Codec;
 import org.opendaylight.yangtools.yang.model.util.type.BaseTypes;
 
@@ -25,7 +26,7 @@ public class Uint32CodecStringTest {
     @Test
     public void testSerialize() {
         Uint32Codec<String> codec = getCodec(BaseTypes.uint32Type(), Uint32Codec.class);
-        assertEquals("serialize", "10", codec.serialize(Long.valueOf(10)));
+        assertEquals("serialize", "10", codec.serialize(Uint32.valueOf(10)));
     }
 
     @SuppressWarnings("unchecked")
